--- frontend/src/Components/Signin/Signin.scss
@@ -1,32 +1,0 @@
-@import "../../../node_modules/bootstrap/scss/bootstrap-grid.scss";
-@import "../../../node_modules/bootstrap/scss/grid";
-@import "../../../node_modules/bootstrap/scss/variables";
-@import "../../../node_modules//bootstrap/scss/mixins";
-
-$signin_theme: #b4dbf1;
-
-.bg-signin-card {
-  background-color: $signin_theme;
-}
-
-.divider {
-  width: 100%;
-  text-align: center;
-  border-bottom: 1px solid #000;
-  line-height: 0.1em;
-  margin: 20px 0 20px;
-}
-
-.text-between-divider {
-  background-color: $signin_theme;
-  padding: 0 10px;
-}
-
-.w-signin {
-  @include media-breakpoint-down(md) {
-    max-width: 300px;
-  }
-  @include media-breakpoint-up(md) {
-    width: 350px;
-  }
-}
