+++ frontend/src/pages/SignPages/SignPages.scss
@@ -1,5 +1,20 @@
-@import '../../scss/variables';
-.sign-page-background{
-    background: linear-gradient(180deg, rgba(230, 231, 230, 0.40) 47.57%, rgba(221, 206, 206, 0.00) 100%), $primary;
-    min-height: 100vh;
+@import "../../scss/variables";
+
+.sign-page-container {
+  display: flex;
+  align-items: stretch;
+  flex-grow: 1;
+  padding: 0 !important;
+}
+
+.sign-page-logo-section,
+.sign-page-sign-section {
+  padding: 30px 0;
+}
+
+.sign-page-sign-section {
+  background: linear-gradient(135deg, #e9e1c6, #b2d3fa);
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
