+++ frontend/src/pages/SignPages/Signup.jsx
@@ -1,26 +1,20 @@
-import { Col, Container, Row, Stack } from "react-bootstrap";
-import SignupCard from "../../Components/Signup/SignupCard";
-import Logo from "../../Components/Signin_Signup/Logo";
-import './SignPages.scss';
+import { Col, Container, Row } from "react-bootstrap";
+import SignupCard from "../../Components/Signin_Signup/Signup/SignupCard";
+import Logo from "../../Components/Signin_Signup/Logo/Logo";
+import "./SignPages.scss";
 export default function Signup() {
   return (
     <>
-      <Stack className="sign-page-background py-5 d-flex align-items-center justify-content-center">
-        <Container
-          fluid
-          className="d-flex align-items-center"
-          style={{ maxWidth: "1000px" }}
-        >
-          <Row className="w-100 gy-5 m-0">
-            <Col xs={12} sm={6}>
-              <Logo />
-            </Col>
-            <Col xs={12} sm={6}>
-              <SignupCard />
-            </Col>
-          </Row>
-        </Container>
-      </Stack>
+      <Container fluid className="sign-page-container">
+        <Row className="w-100 gy-5 m-0">
+          <Col xs={12} sm={6} className="sign-page-logo-section m-0 ">
+            <Logo />
+          </Col>
+          <Col xs={12} sm={6} className="m-0 sign-page-sign-section">
+            <SignupCard />
+          </Col>
+        </Row>
+      </Container>
     </>
   );
 }
