+++ frontend/src/scss/variables/_variables-universal.scss
@@ -13,6 +13,11 @@ $popover-body-bg: #2c6ae1;
 $settings-input-bg: #d5d5d5;
 $settings-article-search-bar: #d5d5d5;
 
+//Sign Pages
+$sign-page-title: #366092;
+$welcome-section-link: #4E8AD1;
+$sign-input-placeholder: #6B6565;
+
 //fonts
 @font-face {
   font-family: "Fredericka The Great";
